// packages/playwright-storybook/src/types.ts
/**
 * Copyright (c) Microsoft Corporation.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type StoryEntry = {
  id: string;
  title: string;
  name: string;
  importPath?: string;
  type: 'story' | 'docs';
  tags?: string[];
};

export type StoryIndex = {
  v: number;
  entries: Record<string, StoryEntry>;
};

export type DiscoveryOptions = {
  url?: string;
  include?: string | string[];
  exclude?: string | string[];
  tags?: { include?: string[]; exclude?: string[] };
  timeoutMs?: number;
};